+++ packages/token-metadata/src/TokenFactory.ts
@@ -4,10 +4,10 @@ import { INJ_DENOM } from '@injectivelabs/utils'
 import { TokenInfo } from './TokenInfo'
 import { TokenMetaUtils } from './TokenMetaUtils'
 import {
-  tokensBySymbolForDevnet,
-  tokensBySymbolForDevnet1,
-  tokensBySymbolForDevnet2,
-  tokensBySymbolForTestnet,
+  getTokensBySymbolForDevnet,
+  getTokensBySymbolForDevnet1,
+  getTokensBySymbolForDevnet2,
+  getTokensBySymbolForTestnet,
 } from './tokens/network'
 import { Token, TokenMeta, TokenType } from './types'
 import tokensBySymbol from './tokens/tokens'
@@ -22,21 +22,30 @@ export class TokenFactory {
 
   static make(network: Network = Network.Mainnet): TokenFactory {
     switch (network) {
+      case Network.Staging:
       case Network.Mainnet:
       case Network.MainnetK8s:
       case Network.MainnetLB:
       case Network.Local:
         return new TokenFactory(new TokenMetaUtils(tokensBySymbol))
       case Network.Devnet:
-        return new TokenFactory(new TokenMetaUtils(tokensBySymbolForDevnet))
+        return new TokenFactory(
+          new TokenMetaUtils(getTokensBySymbolForDevnet()),
+        )
       case Network.Devnet1:
-        return new TokenFactory(new TokenMetaUtils(tokensBySymbolForDevnet1))
+        return new TokenFactory(
+          new TokenMetaUtils(getTokensBySymbolForDevnet1()),
+        )
       case Network.Devnet2:
-        return new TokenFactory(new TokenMetaUtils(tokensBySymbolForDevnet2))
+        return new TokenFactory(
+          new TokenMetaUtils(getTokensBySymbolForDevnet2()),
+        )
       case Network.Testnet:
       case Network.TestnetOld:
       case Network.TestnetK8s:
-        return new TokenFactory(new TokenMetaUtils(tokensBySymbolForTestnet))
+        return new TokenFactory(
+          new TokenMetaUtils(getTokensBySymbolForTestnet()),
+        )
       default:
         return new TokenFactory(new TokenMetaUtils(tokensBySymbol))
     }