+++ packages/ts-types/src/index.ts
@@ -22,12 +22,3 @@ export interface PaginationOption {
 export interface Constructable<T> {
   new (...args: never): T
 }
-
-export type Web3GatewayMessage<T> = T & { '@type': string }
-
-export interface ComposerResponse<T, R> {
-  web3GatewayMessage: Web3GatewayMessage<R> | Web3GatewayMessage<R>[]
-  directBroadcastMessage:
-    | { type: string; message: T }
-    | { type: string; message: T }[]
-}