+++ packages/exceptions/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.59](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exceptions@1.0.58...@injectivelabs/exceptions@1.0.59) (2023-03-05)
+
+**Note:** Version bump only for package @injectivelabs/exceptions
+
+
+
+
+
+## [1.0.58](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exceptions@1.0.57...@injectivelabs/exceptions@1.0.58) (2023-03-01)
+
+**Note:** Version bump only for package @injectivelabs/exceptions
+
+
+
+
+
+## [1.0.57](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exceptions@1.0.56...@injectivelabs/exceptions@1.0.57) (2023-03-01)
+
+**Note:** Version bump only for package @injectivelabs/exceptions
+
+
+
+
+
 ## [1.0.56](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/exceptions@1.0.56-alpha.0...@injectivelabs/exceptions@1.0.56) (2023-02-20)
 
 **Note:** Version bump only for package @injectivelabs/exceptions